Move Authenticator and Session to util folder
